cpp
Answer would be true if we can reach to last stone using some number of viable jump size, and false if we could not reach to last stone with any number of viable jump size.

Algorithm:

Make a hashmap of each stone, which would contain a set. That set would store the viable jump sizes that could be used to reach at that stone. E.g.: For 2nd stone, it would be 1 only.
Traverse the stone array from 1st index (2nd stone). That can be the current position of frog at any time.
Populate the sets of stones that can be reached from current position using a viable jump size (that is k - 1 or k or k + 1). Next Viable jump size can be calculated from previous jump size, and this previous jump size would be found in set of current position.
Return true if set of last stone is unempty, that means we can reach to it somehow, false otherwise.